Glucose in Serum

Dataset corresponding to serum glucose (measurements of glucose concentration in blood used to control diabetes) testing. Eight laboratories conducted tests to five different blood samples tagged with different references, ranging them from low sugar content to very high. Three replicates were obtained for each sample. It is retrieved from ASTM E 691 standard.

Format

A data frame with 120 observations composed of the following 4 variables:

Glucose

Glucose content in Serum.

Replicate

Number of glucose measurement corresponding to each material.

Material

Level of glucose, ranging from low content of sugar to very high level of glucose in blood.

Laboratory

Laboratories conducted tests.

References

ASTM E 691 (1999). Standard practice for conducting an interlaboratory study to determine the precision of a test method. American Society for Testing and Materials. West Conshohocken, PA, USA.

Examples

library(ILS)
data(Glucose)
summary(Glucose)
#>     Glucose         Replicate   Material          Laboratory       
#>  Min.   : 39.02   Min.   :1   Length:120         Length:120        
#>  1st Qu.: 78.45   1st Qu.:1   Class :character   Class :character  
#>  Median :135.03   Median :2   Mode  :character   Mode  :character  
#>  Mean   :149.09   Mean   :2                                        
#>  3rd Qu.:196.66   3rd Qu.:3                                        
#>  Max.   :309.40   Max.   :3                                        
attach(Glucose)
#> The following object is masked _by_ .GlobalEnv:
#> 
#>     Glucose
#> The following object is masked from package:ILS:
#> 
#>     Glucose
#> The following object is masked from package:nlme:
#> 
#>     Glucose
str(Glucose)
#> 'data.frame':	120 obs. of  4 variables:
#>  $ Glucose   : num  41 41.5 41.4 41.2 42 ...
#>  $ Replicate : num  1 2 3 1 2 3 1 2 3 1 ...
#>  $ Material  : chr  "A" "A" "A" "A" ...
#>  $ Laboratory: chr  "Lab1" "Lab1" "Lab1" "Lab2" ...
table(Replicate,Material,Laboratory)
#> , , Laboratory = Lab1
#> 
#>          Material
#> Replicate A B C D E
#>         1 1 1 1 1 1
#>         2 1 1 1 1 1
#>         3 1 1 1 1 1
#> 
#> , , Laboratory = Lab2
#> 
#>          Material
#> Replicate A B C D E
#>         1 1 1 1 1 1
#>         2 1 1 1 1 1
#>         3 1 1 1 1 1
#> 
#> , , Laboratory = Lab3
#> 
#>          Material
#> Replicate A B C D E
#>         1 1 1 1 1 1
#>         2 1 1 1 1 1
#>         3 1 1 1 1 1
#> 
#> , , Laboratory = Lab4
#> 
#>          Material
#> Replicate A B C D E
#>         1 1 1 1 1 1
#>         2 1 1 1 1 1
#>         3 1 1 1 1 1
#> 
#> , , Laboratory = Lab5
#> 
#>          Material
#> Replicate A B C D E
#>         1 1 1 1 1 1
#>         2 1 1 1 1 1
#>         3 1 1 1 1 1
#> 
#> , , Laboratory = Lab6
#> 
#>          Material
#> Replicate A B C D E
#>         1 1 1 1 1 1
#>         2 1 1 1 1 1
#>         3 1 1 1 1 1
#> 
#> , , Laboratory = Lab7
#> 
#>          Material
#> Replicate A B C D E
#>         1 1 1 1 1 1
#>         2 1 1 1 1 1
#>         3 1 1 1 1 1
#> 
#> , , Laboratory = Lab8
#> 
#>          Material
#> Replicate A B C D E
#>         1 1 1 1 1 1
#>         2 1 1 1 1 1
#>         3 1 1 1 1 1
#> 
table(Laboratory,Material)
#>           Material
#> Laboratory A B C D E
#>       Lab1 3 3 3 3 3
#>       Lab2 3 3 3 3 3
#>       Lab3 3 3 3 3 3
#>       Lab4 3 3 3 3 3
#>       Lab5 3 3 3 3 3
#>       Lab6 3 3 3 3 3
#>       Lab7 3 3 3 3 3
#>       Lab8 3 3 3 3 3
st <- with(Glucose, tapply(Glucose, list(Material,Laboratory), mean))
st
#>        Lab1      Lab2      Lab3      Lab4      Lab5      Lab6      Lab7
#> A  41.28333  41.44000  41.45000  41.45667  41.46333  42.02000  40.45667
#> B  78.31667  79.23333  79.90333  80.96333  78.69000  79.89333  79.51667
#> C 133.19667 135.40667 134.59000 140.83000 133.26667 136.61667 132.49333
#> D 193.65000 195.10667 192.09000 197.21333 193.05000 197.24333 191.26000
#> E 293.25333 298.91667 292.67000 295.82000 293.56333 294.95667 290.13667
#>        Lab8
#> A  42.57667
#> B  80.34667
#> C 134.71000
#> D 198.12333
#> E 296.62000