Dataset composed of the initial decomposition temperature (IDT) of different samples of Calcium Oxalate, obtained by 7 different laboratories

Initial decomposition temperature (IDT) is a parameter defined by temperature at which a material loss 5% of its weight when it is heated using a constant rate. One hundred and five calcium oxalate samples were tested by thermogravimetric analysis (TG), obtaining 105 TG curves from which the IDT is extracted. Summarizing, IDT dataset is composed of the IDT values of calcium oxalate obtained by 7 different laboratories that analyze 15 oxalate samples each one.: Laboratory 1 uses a simultaneous thermal analyzer (STA) with an old calibration program, Laboratory 2 to Laboratory 4 use a SDT simultaneous analyzer, Laboratory 6 utilizes a SDT simultaneous analyzer with an old calibration, and Laboratory 7 uses a SDT simultaneous analyzer with a biased calibration (2 degrees Celsius shifted from the zinc melting point).

Format

Dataframe of dimension 105 x 44. The first column corresponds to IDT variable, the second (Sample) is the replicate number, the third is the tested material (Material), and fourth is the laboratory.

References

Naya, S., Tarrio-Saavedra. J., Lopez- Beceiro, J., Francisco Fernandez, M., Flores, M. and Artiaga, R. (2014), "Statistical functional approach for interlaboratory studies with thermal data". Journal of Thermal Analysis and Calorimetry, 118,1229-1243.

Examples

library(ILS)
data(IDT)
summary(IDT)
#>       IDT            Sample        Run    Laboratory
#>  Min.   :162.4   Min.   : 1   Min.   :1   Lab 1:15  
#>  1st Qu.:164.1   1st Qu.: 4   1st Qu.:1   Lab 2:15  
#>  Median :164.1   Median : 8   Median :1   Lab 3:15  
#>  Mean   :164.4   Mean   : 8   Mean   :1   Lab 4:15  
#>  3rd Qu.:164.9   3rd Qu.:12   3rd Qu.:1   Lab 5:15  
#>  Max.   :166.5   Max.   :15   Max.   :1   Lab 6:15  
#>                                           Lab 7:15  
attach(IDT)
#> The following object is masked _by_ .GlobalEnv:
#> 
#>     IDT
#> The following object is masked from Glucose:
#> 
#>     Laboratory
#> The following object is masked from package:ILS:
#> 
#>     IDT
str(IDT)
#> 'data.frame':	105 obs. of  4 variables:
#>  $ IDT       : num  165 166 165 166 165 ...
#>  $ Sample    : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ Run       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ Laboratory: Factor w/ 7 levels "Lab 1","Lab 2",..: 1 1 1 1 1 1 1 1 1 1 ...
table(Sample,Run,Laboratory)
#> , , Laboratory = Lab 1
#> 
#>       Run
#> Sample 1
#>     1  1
#>     2  1
#>     3  1
#>     4  1
#>     5  1
#>     6  1
#>     7  1
#>     8  1
#>     9  1
#>     10 1
#>     11 1
#>     12 1
#>     13 1
#>     14 1
#>     15 1
#> 
#> , , Laboratory = Lab 2
#> 
#>       Run
#> Sample 1
#>     1  1
#>     2  1
#>     3  1
#>     4  1
#>     5  1
#>     6  1
#>     7  1
#>     8  1
#>     9  1
#>     10 1
#>     11 1
#>     12 1
#>     13 1
#>     14 1
#>     15 1
#> 
#> , , Laboratory = Lab 3
#> 
#>       Run
#> Sample 1
#>     1  1
#>     2  1
#>     3  1
#>     4  1
#>     5  1
#>     6  1
#>     7  1
#>     8  1
#>     9  1
#>     10 1
#>     11 1
#>     12 1
#>     13 1
#>     14 1
#>     15 1
#> 
#> , , Laboratory = Lab 4
#> 
#>       Run
#> Sample 1
#>     1  1
#>     2  1
#>     3  1
#>     4  1
#>     5  1
#>     6  1
#>     7  1
#>     8  1
#>     9  1
#>     10 1
#>     11 1
#>     12 1
#>     13 1
#>     14 1
#>     15 1
#> 
#> , , Laboratory = Lab 5
#> 
#>       Run
#> Sample 1
#>     1  1
#>     2  1
#>     3  1
#>     4  1
#>     5  1
#>     6  1
#>     7  1
#>     8  1
#>     9  1
#>     10 1
#>     11 1
#>     12 1
#>     13 1
#>     14 1
#>     15 1
#> 
#> , , Laboratory = Lab 6
#> 
#>       Run
#> Sample 1
#>     1  1
#>     2  1
#>     3  1
#>     4  1
#>     5  1
#>     6  1
#>     7  1
#>     8  1
#>     9  1
#>     10 1
#>     11 1
#>     12 1
#>     13 1
#>     14 1
#>     15 1
#> 
#> , , Laboratory = Lab 7
#> 
#>       Run
#> Sample 1
#>     1  1
#>     2  1
#>     3  1
#>     4  1
#>     5  1
#>     6  1
#>     7  1
#>     8  1
#>     9  1
#>     10 1
#>     11 1
#>     12 1
#>     13 1
#>     14 1
#>     15 1
#> 
table(Laboratory,Run)
#>           Run
#> Laboratory  1
#>      Lab 1 15
#>      Lab 2 15
#>      Lab 3 15
#>      Lab 4 15
#>      Lab 5 15
#>      Lab 6 15
#>      Lab 7 15
st <- with(IDT, tapply(IDT, list(Run,Laboratory), mean))
st
#>      Lab 1    Lab 2    Lab 3    Lab 4    Lab 5    Lab 6    Lab 7
#> 1 164.4324 164.1081 164.2703 164.1622 164.1622 164.2162 165.7297