Descriptive measures for functional data.

Central and dispersion measures for functional data.

Usage

func.ils.formula(formula, data = NULL, drop = FALSE, func = func.mean)

Arguments

formula

A formula, such as y ~ group, where y is a fdata object. to be split into groups according to the grouping variable group (usually a factor).

data

List that containing the variables in the formula. The item called "df" is a data frame with the grouping variable. The item called "y" is a fdata object.

drop

Logical indicating if levels that do not occur should be dropped (if f is a factor or a list).

func

Measures for functional data.