Quality Control Data

It creates a 'lab.qcdata' class object to perform the interlaboratory study. This object is used to plot ILS data and more.

Usage

lab.qcdata(
  data,
  var.index = 1,
  replicate.index = 2,
  material.index = 3,
  laboratory.index = 4,
  data.name = NULL
)

Arguments

data

A matrix or data-frame that contains the data, replicate index, type of material, and the laboratory.

var.index

A scalar with the column number corresponding to the observed variable (the critical to quality variable). Alternativelly can be a string with the name of the quality variable.

replicate.index

A scalar with the column number corresponding to the index each replicate.

material.index

A scalar corresponding to the replicated number.

laboratory.index

A scalar that defines the index number of each laboratory.

data.name

A string specifying the name of the variable which appears on the plots. If name is not provided, it is taken from the object given as data.

Examples

library(ILS)
data(Glucose)
Glucose.qcdata <- lab.qcdata(Glucose)
str(Glucose.qcdata)
#> Classes ‘lab.qcdata’ and 'data.frame':	120 obs. of  4 variables:
#>  $ x         : num  41 41.5 41.4 41.2 42 ...
#>  $ replicate : Factor w/ 3 levels "1","2","3": 1 2 3 1 2 3 1 2 3 1 ...
#>  $ material  : Factor w/ 5 levels "A","B","C","D",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ laboratory: Factor w/ 8 levels "Lab1","Lab2",..: 1 1 1 2 2 2 3 3 3 4 ...
#>  - attr(*, "data.name")= chr "Glucose"
summary(Glucose.qcdata)
#>        x          replicate material   laboratory
#>  Min.   : 39.02   1:40      A:24     Lab1   :15  
#>  1st Qu.: 78.45   2:40      B:24     Lab2   :15  
#>  Median :135.03   3:40      C:24     Lab3   :15  
#>  Mean   :149.09             D:24     Lab4   :15  
#>  3rd Qu.:196.66             E:24     Lab5   :15  
#>  Max.   :309.40                      Lab6   :15  
#>                                      (Other):30